move InitMask to its own module
